+++ pyzoo/zoo/tfpark/tf_predictor.py
@@ -18,6 +18,8 @@ import sys
 
 from zoo.pipeline.api.net.utils import find_placeholders, _check_the_same
 from zoo.tfpark.tfnet import TFNet
+from zoo.tfpark.tf_dataset import TFNdarrayDataset
+from zoo.tfpark.utils import _standarize_feature_dataset
 
 if sys.version >= '3':
     long = int
@@ -68,6 +70,10 @@ class TFPredictor:
 
         outputs = keras_model.outputs
         inputs = keras_model.inputs
+
+        if isinstance(dataset, TFNdarrayDataset):
+            dataset = _standarize_feature_dataset(dataset, keras_model)
+
         return cls(sess, outputs, inputs, dataset)
 
     def predict(self):
